Correction to DAC_get_data_register

+++ src/dac.h
@@ -9,7 +9,7 @@
 
 void 				DAC_init 				(void);
 void 				DAC_write				(uint16_t u16_data);
-const uint32_t * 	DAC_get_data_register	(void);
+const uint16_t * 	DAC_get_data_register	(void);
 
 uint8_t 			DAC_shell_ramp			(uint8_t argc, char ** argv);
 uint8_t 			DAC_shell_write			(uint8_t argc, char ** argv);
